1, modify queue.cs queue writepointer and readpointer type from byte to int; 2. add stack.cs

+++ DiagTool_Luffy/Queue.cs
@@ -15,8 +15,8 @@ namespace DiagTool_Luffy
         }
 
         public int Depth;
-        public byte WritePointer;
-        public byte ReadPointer;
+        public int WritePointer;
+        public int ReadPointer;
         public bool FullFlag;
         public bool EmptyFlag;
         public object[] Datas;
